WIP test Player.vue
