+++ components/RoutineCarousel.vue
@@ -10,7 +10,8 @@
 </template>
 
 <script lang="ts" setup>
-import { type Step } from "@/store/routine";
+import { type Step } from "@/store/types";
+
 const props = defineProps(['currentStepIndex', 'playerSteps']);
 
 const currentStep = computed((): Step | undefined => {
