+++ composables/clipHtml.ts
@@ -1,4 +1,5 @@
 export function clipHtml(watchable: any) {
+  if (typeof watch === 'undefined') return;
   watch(watchable, (newValue) => {
     if (newValue) {
       document.documentElement.classList.add('is-clipped');
