+++ tests/setup-tests.js
@@ -1,0 +1,4 @@
+import { config } from '@nuxt/test-utils';
+
+// Mock Nuxt client-side component
+config.stubs['client-only'] = '<div><slot /></div>';
